Remove Sepolia from networks that can be selected
When trying TACo encryption or decryption with Sepolia network selected,
it returns an error: "Error: No contract found for name: Coordinator".

If you select Sepolia, the web3 provider (Metamask) will try to call the
Coordinator contract on Sepolia, but the testnet where this contract
lives is Polygon Amoy, so this make the app to fail.